Short circuited interactive tests symlink for existing files to clean up ERROR messages in test output.
git-svn-id: https://aeon.stsci.edu/ssb/svn/crds_server/trunk@2151 37088bc5-bd3f-4c5b-b9de-119bb1e8f7e3

+++ sources/interactive/tests.py
@@ -155,7 +155,7 @@ class InteractiveBase(TransactionTestCase):
             else:
                 source = os.path.abspath(filename)
             where = lconfig.relocate_file(filename, self.observatory)
-            if link:
+            if link and not os.path.exists(where):
                 with log.error_on_exception("Symlinking", repr(source), "to", repr(where), "failed."):
                     log.info("Symlinking fake file", repr(source), "to", repr(where))
                     os.symlink(source, where)
